Preempt prune/deletion operations

+++ hydra-task/src/main/java/com/addthis/hydra/task/output/tree/PathPrune.java
@@ -27,6 +27,7 @@ import com.addthis.hydra.data.tree.TreeNodeList;
 import com.addthis.hydra.data.tree.prop.DataTime;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.util.concurrent.Runnables;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -67,6 +68,15 @@ public class PathPrune extends PathElement {
     @FieldConfig private int relativeDown = 0;
 
     /**
+     * If true then terminate the pruning process when the job is shutting down.
+     * Default is false. Note that specifying a prune in the "post" section and
+     * setting preempt to true can prevent job pruning from happening if the
+     * job always terminates when its maximum runtime is reached.
+     * Consider specifying a prune in the "pre" section and setting preempt to true.
+     */
+    @FieldConfig private boolean preempt = false;
+
+    /**
      * If non-null then parse the name of each node using the provided
      * <a href="http://joda-time.sourceforge.net/apidocs/org/joda/time/format/DateTimeFormat
      * .html">DateTimeFormat</a>. Default is null.
@@ -87,44 +97,72 @@ public class PathPrune extends PathElement {
     public TreeNodeList getNextNodeList(final TreeMapState state) {
         long now = JitterClock.globalTime();
         DataTreeNode root = state.current();
-
-        findAndPruneChildren(root, now, relativeDown);
+        findAndPruneChildren(state, root, now, relativeDown);
         return TreeMapState.empty();
     }
 
-    public void findAndPruneChildren(final DataTreeNode root, long now, int depth) {
+    public void findAndPruneChildren(final TreeMapState state, final DataTreeNode root, long now, int depth) {
         if (depth == 0) {
-            pruneChildren(root, now);
+            pruneChildren(state, root, now);
         } else {
             ClosableIterator<DataTreeNode> keyNodeItr = root.getIterator();
-            while (keyNodeItr.hasNext()) {
-                findAndPruneChildren(keyNodeItr.next(), now, depth - 1);
+            try {
+                while (keyNodeItr.hasNext() && !(preempt && state.processorClosing())) {
+                    findAndPruneChildren(state, keyNodeItr.next(), now, depth - 1);
+                }
+            } finally {
+                keyNodeItr.close();
             }
-            keyNodeItr.close();
         }
     }
 
-    public void pruneChildren(final DataTreeNode root, long now) {
+    public void pruneChildren(final TreeMapState state, final DataTreeNode root, long now) {
         ClosableIterator<DataTreeNode> keyNodeItr = root.getIterator();
         int deleted = 0;
         int kept = 0;
         int total = 0;
-        while (keyNodeItr.hasNext()) {
-            DataTreeNode treeNode = keyNodeItr.next();
-            long nodeTime = getNodeTime(treeNode);
-            if ((nodeTime > 0) && ((now - nodeTime) > ttl)) {
-                root.deleteNode(treeNode.getName());
-                deleted++;
-            } else {
-                kept++;
+        try {
+            if (preempt && (state.processorClosing() || expensiveShutdownTest())) {
+                return;
             }
-            total++;
-            if ((total % 100000) == 0) {
-                logger.info("Iterating through children of {}, deleted: {} kept: {}", root.getName(), deleted, kept);
+            while (keyNodeItr.hasNext() && !(preempt && state.processorClosing())) {
+                DataTreeNode treeNode = keyNodeItr.next();
+                long nodeTime = getNodeTime(treeNode);
+                if ((nodeTime > 0) && ((now - nodeTime) > ttl)) {
+                    root.deleteNode(treeNode.getName());
+                    deleted++;
+                } else {
+                    kept++;
+                }
+                total++;
+                if ((total % 100000) == 0) {
+                    logger.info("Iterating through children of {}, deleted: {} kept: {}", root.getName(), deleted,
+                                kept);
+                }
             }
+        } finally {
+            logger.info("Iterated through children of {}, deleted: {} kept: {}", root.getName(), deleted, kept);
+            keyNodeItr.close();
         }
-        logger.info("Iterated through children of {}, deleted: {} kept: {}", root.getName(), deleted, kept);
-        keyNodeItr.close();
+    }
+
+    /**
+     * It is bad practice use the shutdown hook mechanism as a way
+     * to test whether the JVM is shutting down. However if we use
+     * this method exactly once then it is useful when writing tests
+     * to ensure that zero prune operations occur during shutdown.
+     *
+     * @return true iff the jvm is shutting down
+     */
+    private static boolean expensiveShutdownTest() {
+        try {
+            Thread shutdownHook = new Thread(Runnables.doNothing(), "Path prune shutdown hook");
+            Runtime.getRuntime().addShutdownHook(shutdownHook);
+            Runtime.getRuntime().removeShutdownHook(shutdownHook);
+        } catch (IllegalStateException ignored) {
+            return true;
+        }
+        return false;
     }
 
     @VisibleForTesting long getNodeTime(DataTreeNode treeNode) {
